src/requirements.txt
amqp==2.2.2
anyjson==0.3.3
asgiref==3.4.1
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
billiard==3.5.0.3
bootstrap4==0.1.0
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.2.6
django-bootstrap3==9.1.0
django-bootstrap4==1.1.1
django-celery==3.2.2
django-celery-results==1.0.1
django-crispy-forms==1.9.0
django-debug-panel==0.8.3
django-debug-toolbar==3.2.1
django-debug-toolbar-request-history==0.1.3
django-developer-panel==0.1.2
django-geoip2-extras==1.1.2
django-haystack==2.7.0
django-location-field==2.1.0
django-map-widgets==0.2.2
django-redis==4.7.0
django-tinymce==2.7.0
Faker==8.11.0
geoip2==2.9.0
idna==2.6
image-slicer==0.3.0
importlib-metadata==1.1.0
isort==4.3.21
itsdangerous==0.24
jieba==0.39
kombu==4.1.0
lazy-object-proxy==1.4.3
maxminddb==1.5.1
mccabe==0.6.1
more-itertools==8.0.0
mutagen==1.40.0
mysql-connector-python==8.0.26
mysqlclient @ file:///C:/Users/Ahmed/Desktop/test%20origen/Classified-Ads/ads/src/mysqlclient-1.4.6-cp38-cp38-win32.whl
oauthlib==3.1.0
Pillow==6.2.1
Pillow-PIL==0.1.dev0
pycodestyle==2.5.0
pycryptodomex==3.9.4
pylint==2.4.4
PyMySQL==0.8.0
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyQt5Designer==5.13.1
pytesseract==0.3.0
python-alipay-sdk==1.7.0
python-dateutil==2.8.2
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.9
redis==2.10.6
requests==2.18.4
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.2.4
tesseract==0.1.3
text-unidecode==1.3
urllib3==1.22
vine==1.1.4
virtualenv==16.7.7
whitenoise==5.3.0
Whoosh==2.7.4
wrapt==1.11.2
zipp==0.6.0
